Removes scroll offset from saved bundle during activity restarts.
Up to this point, the app used getFirstVisiblePosition() / setPosition() to recall the rough scroll position and getTop() / scrollBy() to recall the precise scroll position. Using scrollBy() to restore the initial scrolling position of the poster grid view turned out to be problematic. Black areas sometimes appearing on top and the scrolling experience that was not fluid.

I attempted to use an alternative approach by saving the entire grid state with grid.onSaveInstanceState() and grid.onRestoreInstanceState(). It makes sense -- who better to know how to save and restore its internal state than the grid itself? It turned out that this alone does not restore the scroll state, indeed by design. In general, switching between screen orientations with different number of columns are problematic because switching back and forth is not a stable operation.

However, an approximate restore by index only is better than no restore (it's annoying to be thrown back to the top of the list due to an accidental orientation change) so I keep that part, and remove the problematic offset.