Replace FOREIGN KEY ON DELETE CASCADE with triggers
INSERT OR REPLACE works in such a way that a conflict triggering the REPLACE
	cause deletes the conflicting row before a new row is inserted. When used in
	combination with a FOREIGN KEY ON DELETE CASCADE constraint, the
temporary delete affects not just the row being replaced but actually
immediately cascades to all referenced rows as well. In pratice, this means that
when a movie is present in two lists, downloading the second list deletes the
membership of the first list.

Simply setting the foreign key as DEFERRABLE INITALLY DEFERRED,
postponing the constraint until the end of the transaction, is not
sufficient to prevent the cascading delete.

I considered dropping the foreign key constraint completely. It truly
does not give much value to this schema. In particular, it will be
necessary to sweep the database periodically to remove unreferenced
movie data, orphaned when a movie drops out of the lists. However, it
was a challenge to find a solution that works, and triggers seem to do
the job.