Use assets to determine the number of columns in movie discovery
Previously the number of columns were calculated programatically based
on an ideal movie poster width of 1 inch. This worked well as long as
the fragment was displayed full screen and the total width was fixed and
predetermined. When the app runs in dual pane mode on a large device,
the width of the movie discovery grid is not known in advance but
becomes known in the layout phase.

It's possible to listen to layout events and recalculate the new ideal
poster size dynamically. However, I couldn't find a reliable way to
apply the new size to the existing grid layout. On my test device, a
Genymotion Sony Xperia Tablet emulator, the image posters would not respect
the new height value upon rotation of the device.

In an attempt to get around this problem, this commit discards the
manual calculation and replaces it with an integer resource setting
"movielist_columns".

The movie grid still does not properly respond to orientation changes on
the Sony tablet emulator, but it seems to work well on my phone and a
Nexus 7 emulator. Since I achieve the same app behavior with less code,
I decide to keep this implementation.